address pr feedback
- correct spelling mistake
- move s2n_cert_get_info to openssl and rename
- move s2n_cert_info_test to the openssl namespace
- manually indent clang-formatted code
- use s2n_read_test_pem_and_len over alternative
- remove left over header definition
- switch argument ordering for s2n_security_policy
- NULL initialize pointer
- prefer result ensure over result bail
