+++ tls/s2n_security_policies.h
@@ -191,11 +191,11 @@ bool s2n_security_policy_supports_tls13(const struct s2n_security_policy *securi
 int s2n_find_security_policy_from_version(const char *version, const struct s2n_security_policy **security_policy);
 int s2n_validate_kem_preferences(const struct s2n_kem_preferences *kem_preferences, bool pq_kem_extension_required);
 S2N_RESULT s2n_validate_certificate_signature_preferences(const struct s2n_signature_preferences *s2n_certificate_signature_preferences);
-S2N_RESULT s2n_security_policy_get_version(const struct s2n_security_policy *security_policy, const char **version);
-S2N_RESULT s2n_security_policy_validate_certificate(const struct s2n_cert_info *info,
-        const struct s2n_security_policy *security_policy);
+S2N_RESULT s2n_security_policy_get_version(const struct s2n_security_policy *security_policy,
+        const char **version);
 /* Checks to see if a certificate has a signature algorithm that's in our 
  * certificate_signature_preferences list 
  */
-S2N_RESULT s2n_security_policy_validate_sig_scheme_supported(const struct s2n_cert_info *info,
-        const struct s2n_signature_preferences *cert_sig_preferences);
+S2N_RESULT s2n_security_policy_validate_sig_scheme_supported(
+        const struct s2n_signature_preferences *cert_sig_preferences,
+        const struct s2n_cert_info *info);
